[llvm-exegesis] Add explicit error for segmenetation faults
When benchmarking certain snippets that access memory, segmentation
faults are a very real possibility if the memory annotations are not
perfect. These situtations are currently difficult to debug as there is
no indication of which addresses need to be mapped. This patch adds in
support for printing the address a segmentation fault occurred at to
make it easier to gather this information.

Differential Revision: https://reviews.llvm.org/D158052