[LLDB] Update docs on building documentation
This patch updates the documentation to match recent changes and make it more
clear. More specifically, the process for installing sphinx has changed with
the transition to myst with the requirements.txt in llvm/docs being the
preferred method for installation now. In addition, the docs-lldb-html target
is never generated if swig isn't installed, so having something expliti in
the documentation section (even if it is mentioned as a dependency of lldb
itself above) probably doesn't hurt.