[CI] Refactor monolithic-* scripts to use common utils.sh
This patch refactors big chunks of the common functionality shared
between monolithic-linux.sh and monolithic-windows.sh to a separate
script, utils.sh, that is then sourced in both of the files. This makes
it a bit easier to maintain the scripts.

Platform differences should not be a large deal for the setup here as we
are using bash as the shell on both Linux and Windows.

Pull Request: https://github.com/llvm/llvm-project/pull/152199