[lit] Add support for deleting symlinks to directories without -r
Before this change, rm would assume that a symlink to a directory was
actually a directory and require the recursive flag to be passed,
differing from other shells. Given the change in lit is about the same
length as the test change would be (minus tests), I think it makes sense
to just support this in the internal shell.

Pull Request: https://github.com/llvm/llvm-project/pull/158464